moved activities to activity package
