+++ app/src/main/java/com/example/apple/votingapp/activity/SignUpActivity.java
@@ -1,4 +1,4 @@
-package com.example.apple.votingapp;
+package com.example.apple.votingapp.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -11,6 +11,7 @@ import android.widget.EditText;
 import android.widget.ProgressBar;
 import android.widget.Toast;
 
+import com.example.apple.votingapp.R;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
